Fix: Add YSortOffset and fix jumping Z-level
Note: All the calls to data.get() are somewhat ugly, especially because I then have to unwrap them.
I do however not think that there is an elegant solution and i haven't noticed any performance drawback.
Also not sure if that is a concern, I'd just expect it to be.